# static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SCOMMIE",
    "version": "3.0.0",
    "description": "Image manipulation & JSON API, written in Rust + Axum."
  },
  "paths": {
    "/image/badnews": {
      "get": {
        "summary": "Make a bad news meme with your own text",
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "name": "text",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 130
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PNG image",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "string"
                        },
                        "loc": {
                          "type": "string",
                          "nullable": true
                        },
                        "param_type": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "error": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "string"
                        },
                        "loc": {
                          "type": "string",
                          "nullable": true
                        },
                        "param_type": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "error": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "string"
                        },
                        "loc": {
                          "type": "string",
                          "nullable": true
                        },
                        "param_type": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "error": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/image/beautiful": {
      "get": {
        "summary": "Make a beautiful meme using your own image",
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "name": "image",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Image URL"
          }
        ],
        "responses": {
          "200": {
            "description": "PNG image",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "string"
                        },
                        "loc": {
                          "type": "string",
                          "nullable": true
                        },
                        "param_type": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "error": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "string"
                        },
                        "loc": {
                          "type": "string",
                          "nullable": true
                        },
                        "param_type": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "error": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "string"
                        },
                        "loc": {
                          "type": "string",
                          "nullable": true
                        },
                        "param_type": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "error": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/image/blur": {
      "get": {
        "summary": "Apply a blur filter to your image",
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "name": "image",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Image URL"
          },
          {
            "name": "radius",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 2,
              "minimum": 1,
              "maximum": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PNG image",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "string"
                        },
                        "loc": {
                          "type": "string",
                          "nullable": true
                        },
                        "param_type": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "error": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "string"
                        },
                        "loc": {
                          "type": "string",
                          "nullable": true
                        },
                        "param_type": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "error": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "string"
                        },
                        "loc": {
                          "type": "string",
                          "nullable": true
                        },
                        "param_type": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "error": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/image/circle": {
      "get": {
        "summary": "Apply a circle cut to your image",
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "name": "image",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Image URL"
          }
        ],
        "responses": {
          "200": {
            "description": "PNG image",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "string"
                        },
                        "loc": {
                          "type": "string",
                          "nullable": true
                        },
                        "param_type": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "error": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "string"
                        },
                        "loc": {
                          "type": "string",
                          "nullable": true
                        },
                        "param_type": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "error": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "string"
                        },
                        "loc": {
                          "type": "string",
                          "nullable": true
                        },
                        "param_type": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "error": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/image/color": {
      "get": {
        "summary": "Make an image of the color you provided",
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "3 or 6 digit hex color code"
          },
          {
            "name": "width",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 512,
              "minimum": 15,
              "maximum": 1024
            }
          },
          {
            "name": "height",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 512,
              "minimum": 15,
              "maximum": 1024
            }
          },
          {
            "name": "showCode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "radius",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 15,
              "minimum": 0,
              "maximum": 150
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PNG image",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "string"
                        },
                        "loc": {
                          "type": "string",
                          "nullable": true
                        },
                        "param_type": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "error": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "string"
                        },
                        "loc": {
                          "type": "string",
                          "nullable": true
                        },
                        "param_type": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "error": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "string"
                        },
                        "loc": {
                          "type": "string",
                          "nullable": true
                        },
                        "param_type": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "error": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/image/communist": {
      "get": {
        "summary": "Apply a communist filter to your image",
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "name": "image",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Image URL"
          }
        ],
        "responses": {
          "200": {
            "description": "PNG image",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "string"
                        },
                        "loc": {
                          "type": "string",
                          "nullable": true
                        },
                        "param_type": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "error": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "string"
                        },
                        "loc": {
                          "type": "string",
                          "nullable": true
                        },
                        "param_type": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "error": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "string"
                        },
                        "loc": {
                          "type": "string",
                          "nullable": true
                        },
                        "param_type": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "error": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/image/deepfry": {
      "get": {
        "summary": "Apply a deepfry filter to your image",
        "tags": [
          "Files"
        ],
        "parameters": [
          {
            "name": "image",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Image URL"
          },
          {
            "name": "amount",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 2,
              "minimum": 1,
              "maximum": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PNG image",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "string"
                        },
                        "loc": {
                          "type": "string",
                          "nullable": true
                        },
                        "param_type": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "error": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "string"
                        },
                        "loc": {
                          "type": "string",
                          "nullable": true
                        },
                        "param_type": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "error": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "string"
                        },
                        "loc": {
                          "type": "string",
                          "nullable": true
                        },
                        "param_type": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "error": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}